// src/interfaces.d.ts
export type Identifier = string | number;

export interface IMessage<TPayload = any> {
	/** Event or command type */
	type: string;

	aggregateId?: Identifier;
	aggregateVersion?: number;

	sagaId?: Identifier;
	sagaVersion?: number;

	payload?: TPayload;
	context?: any;
}

export type ICommand<TPayload = any> = IMessage<TPayload>;

export type IEvent<TPayload = any> = IMessage<TPayload> & {
	/** Unique event identifier */
	id?: string;
};

export type IEventStream = ReadonlyArray<Readonly<IEvent>>;


/**
 * Minimum aggregate interface, as it's used by default `AggregateCommandHandler`
 */
export interface IAggregate {

	/** Unique aggregate identifier */
	readonly id: Identifier;

	/** Main entry point for aggregate commands */
	handle(command: ICommand): void | Promise<void>;

	/** List of events emitted by Aggregate as a result of handling command(s) */
	readonly changes: IEventStream;

	/** An indicator if aggregate snapshot should be taken */
	readonly shouldTakeSnapshot?: boolean;

	/** Create aggregate snapshot */
	makeSnapshot?(): TSnapshot;
}

export interface IAggregateState {
	schemaVersion?: number;
	constructor: IAggregateStateConstructor;
}

export interface IAggregateStateConstructor extends Function {
	schemaVersion?: number;
	new(): IAggregateState;
}

type TAggregateConstructorParams<TState extends IAggregateState> = {
	/** Unique aggregate identifier */
	id: Identifier,

	/** Aggregate state snapshot, if any */
	snapshot?: TSnapshot<TState>,

	/** Aggregate events, logged after latest snapshot */
	events?: IEventStream,

	/** Aggregate state instance */
	state?: TState
};
// ...
export type TSnapshot<TPayload = object> = {
	/** 
	 * Schema version of the data stored in `state` property.
	 * Snapshots with older schema versions must be passed thru a data migration before applying for a newer schema
	 */
	schemaVersion: string | number;

	/**
	 * Last event that was processed before making a snapshot
	 */
	lastEvent: IEvent;

	/**
	 * Snapshot data
	 */
	data: TPayload;
}